services/terminology/pyproject.toml
[tool.poetry]
name = "terminology"
version = "0.1.0"
description = "FHIR terminology services"
authors = ["Tom Wilson <twilson650@gmail.com>"]
readme = "README.md"
packages = [{include = "terminology", from = "src"}]

[tool.poetry.dependencies]
python = "^3.13"
cql-parser = "^1.0.2"
"fhir.resources" = "^8.1.0"
requests = "^2.32.5"
python-dotenv = "^1.1.1"
cqlpy = "^0.3.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
